+++ src/components/pre-contract/buyer/step5/Step5Environment.vue
@@ -13,8 +13,8 @@
 import { computed, onMounted } from 'vue'
 import { usePreContractStore } from '@/stores/preContract'
 
-import Step5HasPet from './Step5Sub1MoveIn.vue' // true 일때
-import Step5NoPet from './Step5Sub1MoveIncopy.vue'
+import Step5HasPet from './Step5Sub1MoveInPet.vue' // true 일때
+import Step5NoPet from './Step5Sub1MoveIn.vue'
 import Step5Residency from './Step5Sub2Residency.vue'
 
 // Pinia store
